Fix PathBuf parsing in structopt
It appears that the [`FromStr`] implementation for `PathBuf` was only
stabilized in Rust 1.32.0, and tq aims to support 1.31.0 at minimum.

[`FromStr`]: https://doc.rust-lang.org/std/path/struct.PathBuf.html#impl-FromStr

To fix this, we use `parse(from_os_str)` in the `Opt` struct to be as
compatible as possible. As a bonus, this `structopt` option also allows
us to support optional file paths without needing to resort to
`default_value`.